feat : chap4 - 데이터 중심의 영화 예매 시스템 구현
