--- src/main/java/chap3/Customer.java
@@ -1,22 +1,0 @@
-package chap3;
-
-public class Customer {
-    private String name;
-    private String id;
-
-    public Customer(String name, String id) {
-        this.name = name;
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public String toString() {
-        return "Customer{" +
-            "name='" + name + '\'' +
-            '}';
-    }
-}
