--- src/main/java/chap3/DiscountConditionType.java
@@ -1,6 +1,0 @@
-package chap3;
-
-public enum DiscountConditionType {
-    SEQUENCE,
-    PERIOD
-}
