--- src/main/java/chap3/Money.java
@@ -1,50 +1,0 @@
-package chap3;
-
-import java.math.BigDecimal;
-import java.util.Objects;
-
-public class Money {
-    public static final Money ZERO = Money.amountToWon(0);
-
-    private final BigDecimal amount;
-
-    public Money(BigDecimal amount) {
-        this.amount = amount;
-    }
-
-    public static Money amountToWon(long amount) {
-        return new Money(BigDecimal.valueOf(amount));
-    }
-
-    public Money times(double percent) {
-        return new Money(this.amount.multiply(BigDecimal.valueOf(percent)));
-    }
-
-    public Money minus(Money discountAmount) {
-        return new Money(this.amount.subtract(discountAmount.amount));
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Money)) {
-            return false;
-        }
-        Money money = (Money) o;
-        return Objects.equals(amount, money.amount);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(amount);
-    }
-
-    @Override
-    public String toString() {
-        return "Money {" +
-            "amount=" + amount +
-            '}';
-    }
-}
