--- src/main/java/chap3/MovieType.java
@@ -1,7 +1,0 @@
-package chap3;
-
-public enum MovieType {
-    AMOUNT_DISCOUNT,
-    PERCENT_DISCOUNT,
-    NONE_DISCOUNT
-}
