--- src/main/java/chap3/Screening.java
@@ -1,42 +1,0 @@
-package chap3;
-
-import java.time.LocalDateTime;
-
-public class Screening {
-    private Movie movie;
-    private int sequence;
-    private LocalDateTime whenScreened;
-
-    public Screening(Movie movie, int sequence, LocalDateTime whenScreened) {
-        this.movie = movie;
-        this.sequence = sequence;
-        this.whenScreened = whenScreened;
-    }
-
-    public Money calculateFee(int audienceCount) {
-        switch (movie.getMovieType()) {
-            case AMOUNT_DISCOUNT:
-                if(movie.isDiscountable(whenScreened, sequence)) {
-                    return movie.calculateAmountDiscountedFee().times(audienceCount);
-                }
-                break;
-            case PERCENT_DISCOUNT:
-                if(movie.isDiscountable(whenScreened, sequence)) {
-                    return  movie.calculatePercentDiscountedFee().times(audienceCount);
-                }
-                break;
-            case NONE_DISCOUNT:
-                return movie.calculateNoneDiscountedFee().times(audienceCount);
-        }
-        return movie.calculateNoneDiscountedFee().times(audienceCount);
-    }
-
-    @Override
-    public String toString() {
-        return "Screening{" +
-            "movie=" + movie +
-            ", sequence=" + sequence +
-            ", whenScreened=" + whenScreened +
-            '}';
-    }
-}
