+++ src/main/java/chap4/Customer.java
@@ -1,0 +1,11 @@
+package chap4;
+
+public class Customer {
+  private String name;
+  private String id;
+
+  public Customer(String name, String id) {
+    this.name = name;
+    this.id = id;
+  }
+}
