+++ src/main/java/chap4/DiscountCondition.java
@@ -1,0 +1,54 @@
+package chap4;
+
+import java.time.DayOfWeek;
+import java.time.LocalTime;
+
+public class DiscountCondition {
+  private DiscountConditionType type;
+
+  private int sequence;
+
+  private DayOfWeek dayOfWeek;
+  private LocalTime startTime;
+  private LocalTime endTime;
+
+  public DiscountConditionType getType() {
+    return type;
+  }
+
+  public void setType(DiscountConditionType type) {
+    this.type = type;
+  }
+
+  public int getSequence() {
+    return sequence;
+  }
+
+  public void setSequence(int sequence) {
+    this.sequence = sequence;
+  }
+
+  public DayOfWeek getDayOfWeek() {
+    return dayOfWeek;
+  }
+
+  public void setDayOfWeek(DayOfWeek dayOfWeek) {
+    this.dayOfWeek = dayOfWeek;
+  }
+
+  public LocalTime getStartTime() {
+    return startTime;
+  }
+
+  public void setStartTime(LocalTime startTime) {
+    this.startTime = startTime;
+  }
+
+  public LocalTime getEndTime() {
+    return endTime;
+  }
+
+  public void setEndTime(LocalTime endTime) {
+    this.endTime = endTime;
+  }
+}
