+++ src/main/java/chap4/DiscountConditionType.java
@@ -1,0 +1,6 @@
+package chap4;
+
+public enum DiscountConditionType {
+  SEQUENCE,
+  PERIOD
+}
