+++ src/main/java/chap4/Movie.java
@@ -1,0 +1,55 @@
+package chap4;
+
+import java.time.Duration;
+import java.util.List;
+
+public class Movie {
+  private String title;
+  private Duration runningTime;
+  private Money fee;
+  private List<DiscountCondition> discountConditions;
+
+  private MovieType movieType;
+  private Money discountAmount;
+  private double discountPercent;
+
+  public MovieType getMovieType() {
+    return movieType;
+  }
+
+  public void setMovieType(MovieType movieType) {
+    this.movieType = movieType;
+  }
+
+  public Money getFee() {
+    return fee;
+  }
+
+  public void setFee(Money fee) {
+    this.fee = fee;
+  }
+
+  public List<DiscountCondition> getDiscountConditions() {
+    return discountConditions;
+  }
+
+  public void setDiscountConditions(List<DiscountCondition> discountConditions) {
+    this.discountConditions = discountConditions;
+  }
+
+  public Money getDiscountAmount() {
+    return discountAmount;
+  }
+
+  public void setDiscountAmount(Money discountAmount) {
+    this.discountAmount = discountAmount;
+  }
+
+  public double getDiscountPercent() {
+    return discountPercent;
+  }
+
+  public void setDiscountPercent(double discountPercent) {
+    this.discountPercent = discountPercent;
+  }
+}
