+++ src/main/java/chap4/MovieType.java
@@ -1,0 +1,7 @@
+package chap4;
+
+public enum MovieType {
+  AMOUNT_DISCOUNT,
+  PERCENT_DISCOUNT,
+  NONE_DISCOUNT
+}
