+++ src/main/java/chap4/Screening.java
@@ -1,0 +1,33 @@
+package chap4;
+
+import java.time.LocalDateTime;
+
+public class Screening {
+  private Movie movie;
+  private int sequence;
+  private LocalDateTime whenScreened;
+
+  public Movie getMovie() {
+    return movie;
+  }
+
+  public void setMovie(Movie movie) {
+    this.movie = movie;
+  }
+
+  public int getSequence() {
+    return sequence;
+  }
+
+  public void setSequence(int sequence) {
+    this.sequence = sequence;
+  }
+
+  public LocalDateTime getWhenScreened() {
+    return whenScreened;
+  }
+
+  public void setWhenScreened(LocalDateTime whenScreened) {
+    this.whenScreened = whenScreened;
+  }
+}
